+++ package.json
@@ -60,8 +60,8 @@
     "@types/jest": "^23.3.8",
     "@types/node-fetch": "^2.1.2",
     "@types/react": "^16.4.18",
-    "@types/react-easy-chart": "^0.1.4",
     "@types/react-dom": "^16.0.9",
+    "@types/react-easy-chart": "^0.1.4",
     "@types/react-redux": "^6.0.9",
     "@types/react-router-dom": "^4.3.1",
     "@types/recharts": "^1.1.1",
@@ -69,6 +69,8 @@
     "@types/redux": "^3.6.0",
     "@types/redux-devtools-extension": "^2.13.2",
     "@types/redux-thunk": "^2.1.0",
+    "@types/superagent": "^3.8.4",
+    "@types/supertest": "^2.0.6",
     "@types/uuid": "^3.4.4",
     "awesome-typescript-loader": "^5.2.1",
     "babel-core": "^6.26.3",
@@ -86,6 +88,8 @@
     "sass-loader": "^7.1.0",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
+    "superagent": "^4.0.0-beta.5",
+    "supertest": "^3.3.0",
     "ts-jest": "^23.10.4",
     "ts-node": "^7.0.1",
     "typescript": "^3.1.3",
