+++ src/server/controllers/company.ts
@@ -69,8 +69,8 @@ export class CompanyRepository {
       console.log(item);
       if (item.women_managers === 'n/a') item.women_managers = 0;
       if (item.salary_gap === 'n/a') item.salary_gap = 0;
-      this.db.none('UPDATE companies SET full_name = $1, short_name = $2, company_name = $3, description = $4, company_name_possessive = $5, year_founded = $6, number_employees = $7, url = $8, logo = $9, green_buildings = $10, targets_emissions = $11, human_rights = $12, policy_board_diversity = $13, women_managers = $14, nra_score = $15, brady_rating = $16, tsr = $17, salary_gap = $18, community_score = $19 ,charity_amount = $20, yes_manchin = $21, no_manchin = $22, yes_repeal = $23, no_repeal = $24, yes_tax_cut = $25, no_tax_cut = $26, diversity_score = $27, aila_score = $28, fair_score = $29, taxes_paid = $30 WHERE id = $31;', 
-      [item.full_name, item.short_name, item.company_name, item.company_description, item.company_name_possessive, item.year_founded, item.number_employees, item.url, item.logo, item.green_buildings, item.targets_emissions, item.human_rights, item.policy_board_diversity, item.women_managers, item.nra_score, item.brady_rating, item.tsr, item.salary_gap, item.community_score, item.charity_amount, item.yes_manchin, item.no_manchin, item.yes_repeal, item.no_repeal, item.yes_tax_cut, item.no_tax_cut, item.diversity_score, item.aila_score, item.fair_score, item.taxes_paid, item.id])
+      this.db.none('INSERT INTO company_module (id, company_id, company_name, company_name_possessive, green_buildings, targets_emissions, human_rights, policy_board_diversity, women_managers, nra_score, brady_rating, tsr, salary_gap, community_score, charity_amount, yes_manchin, no_manchin, yes_repeal, no_repeal, yes_tax_cut, no_tax_cut, diversity_score, aila_score, fair_score, taxes_paid) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, $20, $21, $22, $23, $24, $25);', 
+      [v4(), item.company_id, item.company_name, item.company_name_possessive, item.green_buildings, item.targets_emissions, item.human_rights, item.policy_board_diversity, item.women_managers, item.nra_score, item.brady_rating, item.tsr, item.salary_gap, item.community_score, item.charity_amount, item.yes_manchin, item.no_manchin, item.yes_repeal, item.no_repeal, item.yes_tax_cut, item.no_tax_cut, item.diversity_score, item.aila_score, item.fair_score, item.taxes_paid])
       .catch((err: any) => {
         console.log(err);
         return 0;
